refactored to reduce repetative code
