+++ dataeng/jobs/createJobsNew.groovy
@@ -1,8 +1,7 @@
 import static analytics.EmrCostReporter.job as EmrCostReporterJob
 import static analytics.ExpireVerticaPassword.job as ExpireVerticaPasswordJob
 import static analytics.SnowflakeExpirePasswords.job as SnowflakeExpirePasswordsJob
-import static analytics.SnowflakeQueueDepth.job as SnowflakeQueueDepthJob
-import static analytics.SnowflakeCreditUsage.job as SnowflakeCreditUsageJob
+import static analytics.SnowflakeCollectMetrics.job as SnowflakeCollectMetricsJob
 import static analytics.DeployCluster.job as DeployClusterJob
 import static analytics.TerminateCluster.job as TerminateClusterJob
 import static analytics.UpdateUsers.job as UpdateUsersJob
@@ -33,7 +32,7 @@ def taskMap = [
     EMR_COST_REPORTER_JOB: EmrCostReporterJob,
     EXPIRE_VERTICA_PASSWORD_JOB: ExpireVerticaPasswordJob,
     SNOWFLAKE_EXPIRE_PASSWORDS_JOB: SnowflakeExpirePasswordsJob,
-    SNOWFLAKE_COLLECT_METRICS_JOB: [SnowflakeQueueDepthJob, SnowflakeCreditUsageJob],
+    SNOWFLAKE_COLLECT_METRICS_JOB: SnowflakeCollectMetricsJob,
     DEPLOY_CLUSTER_JOB: DeployClusterJob,
     TERMINATE_CLUSTER_JOB: TerminateClusterJob,
     UPDATE_USERS_JOB: UpdateUsersJob,
@@ -53,11 +52,7 @@ for (task in taskMap) {
     catch (Exception e) {
         out.println("File $extraVarsFileName does not exist.")
     }
-    if (task.value instanceof List) {
-        task.value.forEach(this, commonConfigMap + extraVars)
-    } else {
-        task.value(this, commonConfigMap + extraVars)
-    }
+    task.value(this, commonConfigMap + extraVars)
 }
 
 listView('Production') {
