+++ dataeng/resources/snowflake-collect-metrics.sh
@@ -5,7 +5,8 @@ set -ex
 cd $WORKSPACE/analytics-tools/snowflake
 make requirements
 
-python $METRIC_COLLECTION_SCRIPT \
+python collect-metrics.py \
+    --metric_name $METRIC_NAME \
     --key_path $WORKSPACE/analytics-secure/snowflake/rsa_key_snowflake_task_automation_user.p8 \
     --passphrase_path $WORKSPACE/analytics-secure/snowflake/rsa_key_passphrase_snowflake_task_automation_user \
     --automation_user $SNOWFLAKE_USER \
